pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hass_mqtt"
version = "0.0.1"
requires-python = ">=3.7"
dependencies = [
    "paho-mqtt",
    "typing_extensions"
]
